implementation de chargement de sequenecs par streaming dans fastaParser

+++ SuffixArray.cpp
@@ -24,13 +24,13 @@ SuffixArray::SuffixArray(const std::string& inputText) : text(inputText + '$') {
 // methode de ma table SA ********************************
 
 void SuffixArray::buildSuffixArray(){
-    int n = text.length(); //ma taille est tjr positif
+    size_t n = text.length(); //ma taille est tjr positif
 
     //declarre un vecteur de pair pour stocker les indice et les siffixes 
-    std::vector<std::pair<std::string, int>> suffixes;
+    std::vector<std::pair<std::string, size_t>> suffixes;
 
     //remplire mon vecteur suffixes 
-    for(int i = 0; i < n; i++){
+    for(size_t i = 0; i < n; i++){
         suffixes.push_back({text.substr(i), i}); // stocker mon suffixe text.substr(i) et ca pos i
     }
 
@@ -45,22 +45,22 @@ void SuffixArray::buildSuffixArray(){
 } 
  //* construire ma table  lcp *****************************
     void SuffixArray::buildLcpArray(){
-        int n =  text.length();  // la taille de mon motif 
+        size_t n =  text.length();  // la taille de mon motif 
         lcpArray.resize(n, 0); // redimentionner ma table lcp avec la taille de mon motif  et initialiser a 0
-        std::vector<int> rank(n, 0); // il va contenir les rangs(L'indice de table suffixarray) des suffixes dans la table SA
+        std::vector<size_t> rank(n, 0); // il va contenir les rangs(L'indice de table suffixarray) des suffixes dans la table SA
 
         //remplir la table rank
-        for(int i = 0; i < n; i++){
+        for(size_t i = 0; i < n; i++){
             rank[suffixArray[i]]= i;
         }
 
-        int k = 0;
-        for(int i=0; i<n; i++){
+        size_t k = 0;
+        for(size_t i=0; i<n; i++){
             if (rank[i] ==  n-1) {
                 k = 0;  //si le rang de est le dernier suffixe alors en reinitialise la longueur de comparaisont k à 0
                 continue;
             }
-            int j = suffixArray[rank[i]+1]; // le suffixe suivant 
+            size_t j = suffixArray[rank[i]+1]; // le suffixe suivant 
             while(i+k <n && j+k <n && text[i+k] == text[j+k]){
                 k++; //incrementre si y a une correspondance 
             }
@@ -73,21 +73,21 @@ void SuffixArray::buildSuffixArray(){
         }
     }
     //methode gitter de SA 
-    const std::vector<int>& SuffixArray::getSuffixArray() const{
+    const std::vector<size_t>& SuffixArray::getSuffixArray() const{
         return suffixArray;
     }
 
     //get ma table lcp
-    const std::vector<int>& SuffixArray::getLcpArray() const{
+    const std::vector<size_t>& SuffixArray::getLcpArray() const{
         return lcpArray;
     }
     
    //chercher un motif dans la table des suffixes (recherghe dichotomique)
 bool SuffixArray::search(const std::string& motif) const{
-    int left = 0; 
-    int right = suffixArray.size() - 1 ; 
-    int res , suffixPos;
-    int mid;
+    size_t left = 0; 
+    size_t right = suffixArray.size() - 1 ; 
+    size_t res , suffixPos;
+    size_t mid;
 
     while(left <= right){
         mid = left + (right - left) / 2;  // calculer le milieu de sa 
@@ -96,7 +96,7 @@ bool SuffixArray::search(const std::string& motif) const{
         std::string suffix = text.substr(suffixPos); // Extraire le suffixe à partir de cette position dans le motif d'origne
         
         // Comparer le motif avec le suffixe
-        int res = motif.compare(0, motif.length(), suffix, 0, motif.length());
+        size_t res = motif.compare(0, motif.length(), suffix, 0, motif.length());
 
         if (res == 0){ /// motif trouvé 
             return true;
@@ -117,16 +117,16 @@ bool SuffixArray::search(const std::string& motif) const{
 //au lieu de comparer les suffixes de la chaîne de caractères
 //cette table est utilisée pour la recherche de motifs dans la chaîne d'origine
 
-bool compareSuffixes(int i, int j, const std::string& s) {
+bool compareSuffixes(size_t i, size_t j, const std::string& s) {
     return s.substr(i) < s.substr(j);
 }
 
 void SuffixArray::buildSuffixArrayEquivalent() {
-    int n = text.length();
-    std::vector<int> indices(n);
+    size_t n = text.length();
+    std::vector<size_t> indices(n);
 
     // Initialiser le vecteur d'indices
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
         indices[i] = i;
     }
 
@@ -139,7 +139,7 @@ void SuffixArray::buildSuffixArrayEquivalent() {
      * au lieu de comparer les suffixes de la chaîne de caractères qui d'une complexité O(n^2logn)   
      * 
      */
-    std::sort(indices.begin(), indices.end(), [this](int i, int j) {
+    std::sort(indices.begin(), indices.end(), [this](size_t i, size_t j) {
         return compareSuffixes(i, j, text);
     });
 
@@ -148,7 +148,7 @@ void SuffixArray::buildSuffixArrayEquivalent() {
 }
 
 //getter de SA equivalent**************
-    const std::vector<int>& SuffixArray::getSuffixArrayEquivalent() const{
+    const std::vector<size_t>& SuffixArray::getSuffixArrayEquivalent() const{
         return SuffixArrayEquivalent;
     }
 
@@ -157,14 +157,14 @@ void SuffixArray::buildSuffixArrayEquivalent() {
  * cette fonction retourne la position du premier suffixe dans la table des suffixes qui est supérieur ou égal à un motif donné.
  * elle est de complexité O(logn)
  */
-int  SuffixArray::lowerBound(const std::string& motif) const{
-    int  left = 0;
-    int  right = suffixArray.size()-1; //-1 pour ne pas calculer le $ a la fin de mon motif
-    int  result = suffixArray.size(); //initialiser le resultat a la taille de ma table SA
+size_t  SuffixArray::lowerBound(const std::string& motif) const{
+    size_t  left = 0;
+    size_t  right = suffixArray.size()-1; //-1 pour ne pas calculer le $ a la fin de mon motif
+    size_t  result = suffixArray.size(); //initialiser le resultat a la taille de ma table SA
     
     while(left <= right){
-        int  mid = left + (right - left) / 2; //calculer le milieu
-        int  cmp =text.compare(suffixArray[mid], motif.length(), motif); //comparer le motif avec le suffixe de milieu
+        size_t  mid = left + (right - left) / 2; //calculer le milieu
+        size_t  cmp =text.compare(suffixArray[mid], motif.length(), motif); //comparer le motif avec le suffixe de milieu
 
         if(cmp >= 0){
             result = mid; //stocker le resultat
@@ -182,14 +182,14 @@ int  SuffixArray::lowerBound(const std::string& motif) const{
  * cette fonction retourne la position du premier suffixe dans la table des suffixes qui est strictement supérieur à un motif donné.
  * elle est de complexité O(logn)
  */
-int SuffixArray::upperBound(const std::string& motif) const{
-        int  left = 0;
-        int  right = suffixArray.size()-1; //-1 pour ne pas calculer le $ a la fin de mon motif
-        int  result = suffixArray.size(); //initialiser le resultat a la taille de ma table SA
+size_t SuffixArray::upperBound(const std::string& motif) const{
+        size_t  left = 0;
+        size_t  right = suffixArray.size()-1; //-1 pour ne pas calculer le $ a la fin de mon motif
+        size_t  result = suffixArray.size(); //initialiser le resultat a la taille de ma table SA
         
         while(left <= right){
-            int  mid = left + (right - left) / 2; //calculer le milieu
-            int  cmp =text.compare(suffixArray[mid], motif.length(), motif); //comparer le motif avec le suffixe de milieu
+            size_t  mid = left + (right - left) / 2; //calculer le milieu
+            size_t  cmp =text.compare(suffixArray[mid], motif.length(), motif); //comparer le motif avec le suffixe de milieu
     
             if(cmp > 0){
                 result = mid; //stocker le resultat
@@ -211,17 +211,17 @@ int SuffixArray::upperBound(const std::string& motif) const{
  * de calculer le nombre d'occurrences du motif dans la chaîne d'origine en utilisant la table des suffixes.
  * la complexité spaciale est de O(1) car on utilise la table des suffixes déja calculer pour calculer le nombre d'occurrences du motif.
  */
-int SuffixArray::countOccurrences(const std::string& text) const {
+size_t SuffixArray::countOccurrences(const std::string& text) const {
     if (text.empty()) {
         return 0;
     }
-    int lower = lowerBound(text);
-    int upper = upperBound(text);
+    size_t lower = lowerBound(text);
+    size_t upper = upperBound(text);
     return upper - lower;
 }
 
 //fonction getfactor
-std::string SuffixArray::getFactor(int i, int k) const {
+std::string SuffixArray::getFactor(size_t i, size_t k) const {
     if (i + k > text.length() - 1) {
         throw std::out_of_range("Index out of range");
     }
